src/requirements.txt
pymongo
amba-event-stream